# lib/swagger/doc/internal.yaml
swagger: '2.0'
info:
  description: Helo Whatsapp API documentation
  version: 1.0.0
  title: Helo Whatsapp
  termsOfService: ''
  contact:
    name: vivaconnect support
    email: support@vivaconnect.co
...
schemes:
  - http
paths:
...
  /audience:
    get:
      isPublic: true
      description: |-
        This api returns list of templates based on filters
        following
        are
        the list of filters: 
        phoneNumber
        firstMessageActivation
        segmentId
        channel
        optinSourceId
        optin
        tempOptin
      summary: Get audience list
      tags:
        - audience
      operationId: GetAudienceRecordList
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: phoneNumber
          in: query
          required: false
          type: string
          description: ''
        - name: firstMessageActivation
          in: query
          required: false
          type: string
          description: ''
        - name: segmentId
          in: query
          required: false
          type: string
          description: ''
        - name: channel
          in: query
          required: false
          type: string
          description: ''
        - name: optinSourceId
          in: query
          required: false
          type: string
          description: ''
        - name: optin
          in: query
          required: false
          type: boolean
          description: ''
        - name: tempOptin
          in: query
          required: false
          type: boolean
          description: ''
      responses:
        '200':
          description: List of all Templates for audience record
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              - code: 2000
                msg: Success
                data:
                  - audienceId: c1444136-9b15-4bdb-86c3-c138f5a3a44c
                    phoneNumber: '919820333891'
                    channel: whatsapp
                    firstMessage: '2020-07-07T14:48:58.251Z'
                    lastMessage: '2020-07-09T13:02:39.018Z'
                    optin: true
                    tempOptin: true
                    optinSource: direct
                    segment_name: Organic Conversation
                    chatFlowId: null
                    name: Ajay
                    email: ajay@yopmail.com
                    gender: Male
                    country: India
                  - audienceId: b70aeadd-4941-4bda-a4bd-fd5d320e454f
                    phoneNumber: '+91842490276'
                    channel: whatsapp
                    firstMessage: '2020-07-05T14:06:05.000Z'
                    lastMessage: null
                    optin: true
                    tempOptin: null
                    optinSource: direct
                    segment_name: Organic Conversation
                    chatFlowId: null
                    name: Raju
                    email: raju@yopmail.com
                    gender: Male
                    country: India
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
    post:
      isPublic: true
      description: |-
        API to add or update audience
        To add audience do not pass audienceID to update template pass audience ID along with parameters to update
      summary: Add update audience
      tags:
        - audience
      operationId: AddUpdateAudienceData
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/AddupdateAudienceRequest'
      responses:
        '200':
          description: Returns Status of audience info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  audienceId: c1444136-9b15-4bdb-86c3-c138f5a3a44c
                  phoneNumber: '919820333891'
                  channel: whatsapp
                  optin: true
                  optinSourceId: 5ad4bfa0-4dd1-433b-a88b-ace8197f8c38
                  segmentId: a969e894-0317-46ee-95cd-8f56d0b34378
                  chatFlowId: null
                  name: Ajay
                  email: ajay@yopmail.com
                  gender: Male
                  country: India
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
  /audience/optin/segment:
    get:
      description: This api returns list of segment
      summary: Get segment list
      tags:
        - audience
      operationId: GetSegmentData
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: List of all Segment for audience record
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: array
            example:
              - code: 2000
                msg: Success
                data:
                  - segmentId: a969e894-0317-46ee-95cd-8f56d0b34378
                    segmentName: Organic Conversation
                  - segmentId: ba7702c0-2b12-4629-b0a3-fb344615eb54
                    segmentName: Custom Segment Name
                  - segmentId: c2769786-243b-41f2-81bb-c6ccf96287db
                    segmentName: test1
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
    post:
      description: |-
        API to add or update segment
        To add segment do not pass segmentId to update segment pass segment ID along with parameters to update
      summary: Add update segment
      tags:
        - audience
      operationId: AddUpdateSegmentData
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/AddupdateSegmentRequest'
      responses:
        '200':
          description: Returns Status of segment info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  segmentName: test02
                  segmentId: f194da3d-0b62-405e-a512-95797f4bcf41
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
  /audience/optin:
    post:
      isPublic: true
      description: |-
        API to update optin
        To update optinSourceId pass optin source
        ID along with parameters to update
        following are the optin source list :
        1. Captured via any form/check-box/t&c - 154e25f4-44cd-4a48-984d-dbdce569d2c1
        2. When received by direct optin message on Whatsapp - 28614254-12c6-4f44-aa3c-20cf9e9589cb
        3. Optin added using API - 7ccc5ca2-4a8e-4a97-9b49-c5e258bbf649
      summary: Update optin satus and add optin source id
      tags:
        - audience
      operationId: markOptinByPhoneNumberAndAddOptinSource
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/UpdateOptinStatusAndAddOptinSourceIdRequest'
      responses:
        '200':
          description: Returns Status of segment info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  optinSourceId: 5ad4bfa0-4dd1-433b-a88b-ace8197f8c38
                  optin: true
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /audience/optout:
    patch:
      description: |-
        API to update optin
         by phone Number
      summary: Update optin satus by phone number
      tags:
        - audience
      operationId: markOptOutByPhoneNumber
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/UpdateOptinStatusRequest'
      responses:
        '200':
          description: Returns Status of segment info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              - code: 2000
                msg: Success
                data:
                  optin: false
              - code: 3000
                msg: No record found.
                data: null
                error: {}
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
definitions:
...
  AddupdateAudienceRequest:
    title: AddupdateaudienceRequest
    example:
      - audienceId: c1444136-9b15-4bdb-86c3-c138f5a3a44c
        phoneNumber: '919820333891'
        channel: whatsapp
        optin: true
        optinSourceId: 5ad4bfa0-4dd1-433b-a88b-ace8197f8c38
        segmentId: a969e894-0317-46ee-95cd-8f56d0b34378
        chatFlowId: null
        name: Ajay
        email: ajay@yopmail.com
        gender: Male
        country: India
    type: array
    items:
      properties:
        audienceId:
          type: string
        phoneNumber:
          type: string
        channel:
          type: string
        optin:
          type: boolean
        optinSourceId:
          type: string
        segmentId:
          type: string
        chatFlowId:
          type: string
        name:
          type: string
        email:
          type: string
        gender:
          type: string
        country:
          type: string
      required:
        - phoneNumber
        - channel
  AddupdateSegmentRequest:
    title: AddupdateSegmentRequest
    example:
      segmentId: f194da3d-0b62-405e-a512-95797f4bcf41
      segmentName: test
    type: object
    properties:
      segmentId:
        type: string
      segmentName:
        type: string
    required:
      - segmentName
...
  UpdateOptinStatusAndAddOptinSourceIdRequest:
    title: UpdateOptinStatusAndAddOptinSourceIdRequest
    example:
      phoneNumber: '9999999999'
      optinSourceId: 5ad4bfa0-4dd1-433b-a88b-ace8197f8c38
    type: object
    properties:
      phoneNumber:
        type: string
      optinSourceId:
        type: string
    required:
      - optinSourceId
      - phoneNumber
  UpdateOptinStatusRequest:
    title: UpdateOptinStatusRequest
    example:
      phoneNumber: '9999999999'
    type: object
    properties:
      phoneNumber:
        type: string
    required:
      - phoneNumber
...